+++ src/jgame/structures/time/InfiniteTimer.java
@@ -1,7 +1,6 @@
 package jgame.structures.time;
 
 import java.util.TimerTask;
-
 import jgame.Behavior;
 
 /**
@@ -11,15 +10,15 @@ import jgame.Behavior;
  */
 public class InfiniteTimer extends Timer{
 
-	private int timeBetweenTasks;
-	
+	private long timeBetweenTasks;
+
 	/**
 	 * 
 	 * @param milli the time between two successive tasks
 	 * @param task the task to execute
 	 * @since 1.0
 	 */
-	public InfiniteTimer(int milli, Behavior task) {
+	public InfiniteTimer(long milli, Behavior task) {
 		super(-1, task);
 		timeBetweenTasks = milli;
 	}
@@ -32,7 +31,10 @@ public class InfiniteTimer extends Timer{
 		task = new TimerTask(){
 			@Override
 			public void run() {
-				timerExecutionTask.run();
+				//synchronized(AnimationsThread.lock){
+					//System.out.println("TIMER");
+					timerExecutionTask.run();
+				//}
 			}	
 		};
 		timer.schedule(task, timeBetweenTasks, timeBetweenTasks);
